Fix parsing to allow CR or LF at buffer end.
This allows Mojo::IRC to deal with dubiously conforming servers that always
emit \n (or \r). This was existing behavior for most of the buffer, just not
the terminating character matching.

This is roughly in line (but a little more leniant than) with behavior of
clients like irssi, which only looks for \n for message delimiting (and
discards \r).

Also update Test::Mojo::IRC pseudo-server to use lenient parsing in accepting
messages (per https://tools.ietf.org/html/rfc2813#section-5 for servers) and
strictly output \r\n (replacing any of \r or \n with \r\n together).